cloud-functions/requirements.txt
functions-framework==3.*
flask==3.*
flask-cors>=4.0.0
google-cloud-firestore>=2.0.0
google-generativeai>=0.8.3

python-dotenv
requests
colorama
